requirements.txt
asgiref==3.4.1
astroid==2.7.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.3.0
psycopg2==2.9.1
PyJWT==1.7.1
pylint==2.10.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
